+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - bug fixes
     - matrix: prevent error when `quantity` is invalid
+    - multixy/xy: send value when setting `spring` to `true` (only if current value is different than spring value)
 
 ## 1.29.5